chore: move lib/util → lib/core/util; graduate request_features_popup to settings
