+++ client-app/lib/core/util/app_version.dart
