+++ client-app/lib/core/util/constants.dart
