+++ client-app/lib/core/util/enums.dart
