+++ client-app/lib/core/util/logger.dart
