+++ client-app/lib/features/auth/view/login_screen.dart
@@ -4,9 +4,9 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:loading_animation_widget/loading_animation_widget.dart';
 import 'package:plan_sync/core/services/theme_service.dart';
 import 'package:plan_sync/features/auth/viewmodel/login_view_model.dart';
-import 'package:plan_sync/util/enums.dart';
-import 'package:plan_sync/util/external_links.dart';
-import 'package:plan_sync/util/snackbar.dart';
+import 'package:plan_sync/core/util/enums.dart';
+import 'package:plan_sync/core/util/external_links.dart';
+import 'package:plan_sync/core/util/snackbar.dart';
 import 'package:provider/provider.dart';
 
 class LoginScreen extends StatelessWidget {
