+++ client-app/lib/features/electives/view/widgets/electives_sem_bar.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:loading_animation_widget/loading_animation_widget.dart';
 import 'package:plan_sync/features/filters/viewmodel/filter_view_model.dart';
-import 'package:plan_sync/util/logger.dart';
-import 'package:plan_sync/util/snackbar.dart';
+import 'package:plan_sync/core/util/logger.dart';
+import 'package:plan_sync/core/util/snackbar.dart';
 import 'package:provider/provider.dart';
 
 class ElectiveSemesterBar extends StatefulWidget {
