+++ client-app/lib/features/settings/view/settings_screen.dart
@@ -3,9 +3,9 @@ import 'package:flutter/services.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:plan_sync/features/settings/viewmodel/settings_view_model.dart';
-import 'package:plan_sync/util/constants.dart';
-import 'package:plan_sync/util/external_links.dart';
-import 'package:plan_sync/util/snackbar.dart';
+import 'package:plan_sync/core/util/constants.dart';
+import 'package:plan_sync/core/util/external_links.dart';
+import 'package:plan_sync/core/util/snackbar.dart';
 import 'package:plan_sync/widgets/bottom-sheets/bottom_sheets_wrapper.dart';
 import 'package:plan_sync/features/settings/view/widgets/logout_button.dart';
 import 'package:plan_sync/widgets/popups/popups_wrapper.dart';
