+++ client-app/lib/widgets/popups/report_error_mail_popup.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:plan_sync/features/filters/viewmodel/filter_view_model.dart';
-import 'package:plan_sync/util/enums.dart';
-import 'package:plan_sync/util/external_links.dart';
-import 'package:plan_sync/util/snackbar.dart';
+import 'package:plan_sync/core/util/enums.dart';
+import 'package:plan_sync/core/util/external_links.dart';
+import 'package:plan_sync/core/util/snackbar.dart';
 import 'package:provider/provider.dart';
 
 class ReportErrorMailPopup extends StatelessWidget {
