+++ client-app/test/mock_controllers/filter_view_model_mock.dart
@@ -2,7 +2,7 @@ import 'package:mockito/mockito.dart';
 import 'package:plan_sync/core/repositories/app_preferences_repository.dart';
 import 'package:plan_sync/features/filters/viewmodel/filter_view_model.dart';
 import 'package:flutter/material.dart';
-import 'package:plan_sync/util/enums.dart';
+import 'package:plan_sync/core/util/enums.dart';
 
 class MockFilterViewModel extends Mock
     with ChangeNotifier
