+++ client-app/test/widgets/time_table_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:plan_sync/util/enums.dart';
+import 'package:plan_sync/core/util/enums.dart';
 import 'package:plan_sync/features/schedule/view/widgets/time_table.dart';
 import 'package:shared_preferences/shared_preferences.dart';
